mapping implementation and original dataset reused
turns out you can't feed the model random numbers and expect actual words as output (mind blowing, right?). so starting the preprocessing pipeline from the very beginning, but this time with the mapping feature added. using these mappings our model will be able to translate the encoded and scaled outputs that it gives us back to our beautiful human language. so the prediction makes sense, instead of looking like a random string of numbers. Also, i've said this before but find it worth mentioning again: the main/original dataset is too big to be committed to github, but i've left a link to the original in my README file. so you can check that out to get a look at what the original dataset looks like.