model retrained and cleaner optimiztion
retrained the model on a cleaned dataset, since it appears that the previous "Sampled_dataset.csv" had some problems. added the mapping format feature to all of the encoding methods for further flexibility. wrote the initial user input pipeline, yet to be optimized and cleaned. also got all the needed mappings for our user input's processing